backend/requirements.txt
Package                      Version
---------------------------- ---------
aiohappyeyeballs             2.6.1
aiohttp                      3.11.16
aiosignal                    1.3.2
annotated-types              0.7.0
anyio                        4.9.0
attrs                        25.3.0
beautifulsoup4               4.13.3
cachetools                   5.5.2
certifi                      2025.1.31
charset-normalizer           3.4.1
dataclasses-json             0.6.7
filetype                     1.2.0
frozenlist                   1.5.0
google-ai-generativelanguage 0.6.17
google-api-core              2.24.2
google-api-python-client     2.166.0
google-auth                  2.38.0
google-auth-httplib2         0.2.0
google-cloud-core            2.4.3
googleapis-common-protos     1.69.2
greenlet                     3.1.1
grpcio                       1.71.0
grpcio-status                1.71.0
h11                          0.14.0
httpcore                     1.0.7
httplib2                     0.22.0
httpx                        0.28.1
httpx-sse                    0.4.0
idna                         3.10
jsonpatch                    1.33
jsonpointer                  3.0.0
langchain                    0.3.23
langchain-community          0.3.21
langchain-core               0.3.51
langchain-google-community   2.0.7
langchain-google-genai       2.1.2
langchain-text-splitters     0.3.8
langsmith                    0.3.24
markdown-it-py               3.0.0
marshmallow                  3.26.1
mdurl                        0.1.2
multidict                    6.3.2
mypy-extensions              1.0.0
numpy                        2.2.4
orjson                       3.10.16
packaging                    24.2
pip                          24.3.1
propcache                    0.3.1
proto-plus                   1.26.1
protobuf                     5.29.4
pyasn1                       0.6.1
pyasn1_modules               0.4.2
pydantic                     2.11.2
pydantic_core                2.33.1
pydantic-settings            2.8.1
Pygments                     2.19.1
pyparsing                    3.2.3
python-dotenv                1.1.0
PyYAML                       6.0.2
requests                     2.32.3
requests-toolbelt            1.0.0
rich                         14.0.0
rsa                          4.9
sniffio                      1.3.1
soupsieve                    2.6
SQLAlchemy                   2.0.40
tenacity                     9.1.2
typing_extensions            4.13.1
typing-inspect               0.9.0
typing-inspection            0.4.0
uritemplate                  4.1.1
urllib3                      2.3.0
yarl                         1.18.3
zstandard                    0.23.0
